backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
motor==3.5.1
pymongo==4.8.0
redis==5.0.8
pinecone-client==4.1.0
google-generativeai==0.8.3
firebase-admin==6.5.0
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
pypdf2==3.0.1
aiofiles==24.1.0
httpx==0.27.0